backend/requirements.txt
fastapi 
uvicorn 
websockets 
redis 
pymongo
twilio 
boto3 
python-dotenv
